+++ src/Service/CloudFormation/src/CloudFormationClient.php
@@ -169,6 +169,14 @@ class CloudFormationClient extends AbstractApi
                     'signService' => 'cloudformation',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://cloudformation.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'cloudformation',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://cloudformation.us-isob-east-1.sc2s.sgov.gov',