+++ src/Service/CloudWatch/src/CloudWatchClient.php
@@ -382,6 +382,14 @@ class CloudWatchClient extends AbstractApi
                     'signService' => 'monitoring',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://monitoring.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'monitoring',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://monitoring.us-isob-east-1.sc2s.sgov.gov',