+++ src/Service/Ecr/src/EcrClient.php
@@ -206,6 +206,14 @@ class EcrClient extends AbstractApi
                     'signService' => 'ecr',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://api.ecr.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'ecr',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://api.ecr.us-isob-east-1.sc2s.sgov.gov',