+++ src/Service/S3/src/S3Client.php
@@ -3323,6 +3323,14 @@ class S3Client extends AbstractApi
                     'signService' => 's3',
                     'signVersions' => ['s3v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://s3.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 's3',
+                    'signVersions' => ['s3v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://s3.us-isob-east-1.sc2s.sgov.gov',