+++ src/Service/Sqs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- AWS api-change: Added `fips-ca-central-1` and `fips-ca-west-1` regions
+
 ## 2.3.0
 
 ### Added