+++ src/Service/Sqs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - AWS api-change: Added `fips-ca-central-1` and `fips-ca-west-1` regions
+- AWS api-change: TODO
 
 ## 2.3.0